requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0
textstat>=0.7.0
nltk>=3.8.0
pydantic>=2.0.0
scikit-learn==1.3.2
openpyxl==3.1.2
